feat: 게시물 UI 컴포넌트 분리 진행 중
